+++ examples/correlation_matrices/GaussianHMC.cpp
