+++ examples/correlation_matrices/Point_matrix.cpp
