+++ examples/correlation_matrices/RandomPointGenerator.cpp
